# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Table:
    properties:
      capacity:
        type: integer
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      location:
        type: string
      table_number:
        type: integer
    type: object
  models.TableInput:
    properties:
      capacity:
        minimum: 1
        type: integer
      is_active:
        type: boolean
      location:
        enum:
        - unutra
        - terasa
        - vip
        type: string
      table_number:
        type: integer
    required:
    - capacity
    - table_number
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend API for Kelic Restaurant
  title: Kelic Restaurant API
  version: "1.0"
paths:
...
  /tables:
    get:
      description: Javni endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Table'
            type: array
      summary: Dohvati sve stolove
      tags:
      - Stolovi
    post:
      consumes:
      - application/json
      parameters:
      - description: Stol
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.TableInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Table'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Dodaj stol (admin)
      tags:
      - Stolovi
  /tables/{id}:
    delete:
      parameters:
      - description: UUID stola
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Obriši stol (admin)
      tags:
      - Stolovi
    put:
      consumes:
      - application/json
      parameters:
      - description: UUID stola
        in: path
        name: id
        required: true
        type: string
      - description: Stol
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.TableInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Table'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Ažuriraj stol (admin)
      tags:
      - Stolovi
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"